Taking a Stance - Verson 2
HTML & CSS landing page. The animate.css file makes items slide into the page and makes it look more dynamic. The Mediaelementplayer.css, styles and gives function to the media bar players. The three style sheets give the whole site styles. toastr is for the form submission and the hero-bg.jpg is for the main background.

The next updates will be the scripts and more images.